add a dockerfile
By using docker[0] we can make it very easy for someone who is interested in
spinning up a new instance to get started. Instead of installing all of the
dependencies and going through the hassle of having the right versions,
requiring only one command like `make docker_run` is a much better experience.

For now the container is cloning the source so it's less focused on using the
local files. This can be changed if there is a strong need for making it easier
to test the local changes, but it would require some more thought.

Some changes were made to the Makefile. The name field is now being set for the
elasticsearch container which makes it easier to use with other containers via
the `--link` option and we have four more targets[1] to the Makefile. The
targets are there primarily so people are not required to remember the options
but can just rely on using `make docker_X` to trigger the actions.

[0]: https://www.docker.com/
[1]:
- `make docker` - Builds the container.
- `make docker_deploy` - Builds the container and uploads it to hub.docker.com.
- `make docker_run` - Builds and then runs the container.
- `make docker_push` - Uploads a built container to hub.docker.com.
Note that uploading to docker hub might require having a account.

This is a work in progress.