Add support for Vagrant
Vagrant[0] makes it easier to create a reproducible environment and can
potentially help new people to get a environment up and running by. This setup
uses the upsteam version of docker and it also adds the vagrant user to the
docker group while not required but for debug nice to have access to docker. The
vm should have everything you need including elasticsearch and various
dependencies required to run the application.

Using this should look something like:

  vagrant up
  vagrant ssh
  cd /vagrant && make run

Later when you have changes:
  vagrant reload

This is also something we could use with some continuous integration.  Not sure
about travis, but Jenkins has a plugin[1] which might be useful.

Please note that docker is not included in install-deps-debian.sh so that the
script is still useful for non docker users.

[0]: https://www.vagrantup.com/
[1]: https://wiki.jenkins-ci.org/display/JENKINS/Vagrant-plugin