Makefile: change project name variable
For some reason I couldn't register hioa-abi. I suspect dashes are not
allowed, but there does not seem to be any information on this from
docker[0]. The project name can be changed, I just wanted to use
something valid to upload to for now.

Also added the full docker hub URL to the output.

[0]: https://github.com/docker/hub-feedback/issues/158