Install: reword and enforce 80-column
While this might not be necessary the long lines make it harder to read IMO.
Splitting the various options for getting up and running is beneficial because
that makes it easier to skimp the document for the path of interest. This
document can be improved and should be seen as work in progress.