Introduces support to handle upload and download of files to be associated with a documentObject.
Uploading/downloading is part of the expected functionality of a Noark core and this implements the upload as described in section 6.1.1.9. Chunked upload/download has not been implemented.

The change introduces some new exceptions and there are obvious changes to the DocumentObject Hateoas controller and Service classes. The application-*.yml files are updated to reflect this change and a pdf file has been introduced into the curl directory. All changes in this commit reflect the functionality required to upload and download files