Changed documentMedium.id (primary key)  to use GenerationType.IDENTITY @GeneratedValue(strategy = GenerationType.IDENTITY
The reason for this is that  GenerationType.AUTO will not allow me to manually insert tuples define in import.sql and then add documentMedium objects later, while IDENTITY does. What happens is that the primary key counter is not incremented on the manual inserts from import.sql and then when we insert normally, the counter is still at 0, but primary key 0 is already used.

 IDENTITY does not have this problem and continues the primary key counter with the right value. This might only be a H2-database issue though.