When retrieving a list, _links should only be returned if the list is empty.
We have interpreted the standard such that the above statement is true when serialising Noark entities. Prior to this _links were being generated at the end of the payload even if the payload wasn't empty. This commit fixes this by removing the line that printed the Hateoas links for an empty list.

CommonUtils, just removed a newline. The extra newline bugged me.