A bit more work on removing the ability to use English metadata  names.
An example is shown below where valid JSON would allow the metadata name "description" to be used (JSON example 1). The core will require a description field to use the Norwegian term "besrkivelse", (JSON example 2)

{
"description" : "A description"
}
(JSON example 1)

{
"beskrivelse" : "A description"
}
(JSON example 2)

The core has had the ability to use English or Norwegian metadata names. This was to support English speaking students interacting with the core. It was pointed out that "One does not simply break a standard in that manner" and after some discussion it makes sense to remove this functionality. There are too many potential problems with having two interpretations of official metadata names.

This commit just removes the constant values that made up the English metadata names have been removed.