Remove remote link dependencies
We are already using npm so moving to local dependencies is pretty
straightforward. Having issues with getting access to the node modules from the
public/ directory I decided it would make it easier to just have it all in web/
Being able to hack on the web application without internet access is also a nice
gain.

Reported-by: pere