Swap out the form-based login mechanism for a JWT one.
This makes it easier to develop GUI. But also a REST service should return
statuses not redirects. So this will hopefully sort this issue.

Changes are the introduction of classes starting with JWT. These have been
integrated [from](https://github.com/szerhusenBC/jwt-spring-security-demo).

This resulted in some additional changes / removals.

 * Some database column names changed.
 * application-* property files needed new data
 * New library dependencies in pom
 * run_curl.sh updated to the JWT method
 * Login controller added