Added etags to most methods that return a single entity (not list). Not added to any of the ny-* methods. During implementation of this we saw an issue with Record / BasicRecord / RegistryEntry and File / CaseFile.
It's a little unclear if the Repository / Service class should be extending the base or just be an implementation
on its own. This is something we need to look at again, but for the moment it seems OK to allow them all be their
own implementation ... but that may result in duplicated code in the project.

So because of this there are som additional changes at the service / repository level.

During the round of changes, I also added recommended method of dependency injection. Constructor injection
rather than @Autowired.

Also fixed some typos.