Set value of version to 0, instead of letting it be null.
When importing required metadata entities (e.g. dokumentmedium,
arkivstatus), prior to this update, the value for version is null. We
haven't got as far as supporting PUT/PATCH on these entities so it's a
potential bug that has been avoided.

This means we really should set @notnull on this column