Better support for the metadata entities.
This change builds further on the @MappedSuperClass that most metadata entities
can inherit from. This greatly reduces the amount of boilerplate code and
increases maintainability and readability. Effectively, the only real
code that has to be written now is the individual metadata endpoints,
coding up the repositories will be quick. It also shows a rough
template for handling the controller / service / repository and what
needs to be done for the other metadata entities.

DocumentMedium and FondsStatus endpoints should now be finished.