Fix an anti-pattern in the code before it becomes too big and unmaintainable.
When creating a hateoas link a method called getEntityType was called.
This then called a helper method in CommonUtils. This allowed us to
easier implement the hatoeas handler to populate the type of entity that
 was being created.

This anti-pattern was introduced because the codebase evolved to handle
hateoas links and at the time, it seemed like a good solution. When
reading the following blog  [post](https://zeroturnaround.com/rebellabs/watch-out-for-these-10-common-pitfalls-of-experienced-java-developers-architects/)
it became clear that we had introduced this anti-pattern
(pitfall number 8).

While it wasn't a major issue and had negligible effect on
performance, it made the codebase harder to maintain. Each time a
new entity is introduced, we have to hunt down places like CommonUtils
and add the class occurrence to the list. Adding a method,
getEntityBaseType to the interface INoarkSystemIdEntity forces all
entities to identify the entity type they are. This simplifies the code.

This fixes the issue described [here] (https://github.com/HiOA-ABI/nikita-noark5-core/issues/68)

It was a simple fix on entities, but hateoas objects containing links
was not as simple. So when creating a a hateoas list of objects, the
core now requires that you identify the type of entity in the list.

For metadata objects that use shared code the identification is straight
forward but the existing arkivstruktur entities have this hard-wired into
them.