Adding support for discoverability. The core will now return an Allow header for each request detailing which HTTP methods the request supports.
For example. A request to:

   http://localhost:8092/noark5v4/hateoas-api/arkivstruktur/ny-arkivskaper

using any valid method should return the following Allow string

   Allow: POST,GET,HEAD

This is being introduced to enable discoverability and is hinted at in
 the interface.

The big change here is that all Controller methods were updated to
support this. After the application starts it queries all endpoints and
creates a map of endpoint address to supported methods. This can be
found in AfterApplicationStartup and CommonUtilss.

Some small other bugs were fixed, 201 being returned instead of 200.
Precedence stuff is introduced in this push aswell