Fix issues related to eTag.
Outgoing eTags have an extra set of quotes on them. This is because
the MappingJackson2HttpMessageConverter is converting string literals
to JSON and the eTag ends up being ETAG : ""0"".

 https://stackoverflow.com/questions/14293469/spring-mvc-handler-returns-string-with-extra-quotes

explains this an recommends instantiating a StringHttpMessageConverter so
that the eTag is treated as a string and not JSON

Other changes here.

@EnableWebMvc annotation was removed

This was done based on a recommendation by

https://techblog.bozho.net/spring-boot-enablewebmvc-common-use-cases/

and is part of a set of tasks to be carried out in

https://github.com/HiOA-ABI/nikita-noark5-core/issues/75

We have also removed

  configurer.defaultContentType(MediaType.APPLICATION_JSON_UTF8);

as the client must set Accepts. This might come back later.