Fix problem where Content-Type was set to json
We were using a ResponseEntity return type and the default return type
was set to json. This resulted in the return type being set to json as
well for all file downloads. We believe Spring was overwriting the
Content-Type header field. This update fixes this issue and the
Content-Type header field will now be set to whatever the mimeType field
is set to.