Big refactoring job to make code more maintainable and to solve some noted issues.
The big thing here is the introduction of NoarkEntity and
NoarkGeneralEntity and how all entities now inherit from one of these.
This allows us to move forward adding / removing common functionality
from the base classes and makes things more maintainable. At the same
time this refactoring should solve the following issues:

https://github.com/HiOA-ABI/nikita-noark5-core/issues/79
https://github.com/HiOA-ABI/nikita-noark5-core/issues/70 (partly)

It's also a step to help to solve:
https://github.com/HiOA-ABI/nikita-noark5-core/issues/77

Changes to the domain model have a knock on effect into controllers etc,
thus making the push look bigger than it is.