More support for ETAG and introducing support for DELETE
This update disables soft-delete as it is causing a problem with
referential integrity.

The work here sows that there are problems in the domain model. This
became clear when trying to delete. Also when looking at debug output
of queries for delete, it appears that a delete of e.g a record is
deleting up to fonds level recursively.

To be able to handle updates the deserialising method had to be slightly
changed as an updated may contain elements that are not in a POST
request.