Fix problems with RELS in registryEnrty / journalpost
RELS user journalpost where it should use registrering and journalpost
where it should use basisregistrering. This was because the code asks
the entity what type of entity it is. But as discussed in the issue we
agree that we will use the base types rather than the super type.

This is related to
https://github.com/HiOA-ABI/nikita-noark5-core/issues/83