Finalised the PUT functionality in all endpoints.
Only mappe and saksmappe still have a problem with PUT. This is because
mappe contains a referanseArkivdel, referenceSeries element of type
Series, while the incoming payload is a systemID of the series.

file and series are connected via primary keys, not systemID. So this is
one we have to think about. It is also applicable for class.createFile
as well as we have to identify the Series.