Experimental fix for the update of File and CaseFile.
File and CaseFile have a field referanseArkivdel that comes with a
systemID to the Series. I have been unsure how this should be handled,
and am testing out an approach to simply deseiralise it and add it to the
referenceSeries and see if hibernate can sort it out. It appears that
hibernate does this. The reference is kept intact and my fear that
an empty incoming series object created as part of the deserialisation
process would null out the other values was not true.

We need to test this a little further, but it looks OK