Change upload code to verify dokumentobjekt values matches uploaded file.
Changed creation of dokumentobjekt to verify the value of
sjekksumAlgoritme matches the algoritm supported by the API.

Changed file upload to compare the dokumentobjekt values for sjekksum
and filstoerrelse with the uploaded file, and reject the upload if
they do not match.

Changed file upload to only allow one successful upload.  The second
upload will be rejected.

If sjekksum, sjekksumAlgoritme and filstoerrelse is unset when
creating dokumentobjekt, they will get values when a file is
successfully uploaded.