Update README.md
Added clarification about license. The developers have chosen to stick with AGPL rather than APL. Nikita starts its life as teaching and research tool.  If others decide to use nikita then it should be seen as a shared common component (no:felleskomponent). There is no reason why additions to the core should not be made available to others. Choosing AGPL is done to secure this.