Changes in profile handling to make profile handling clearer and more sensible.
ccscanf made a point that profile use should be limited in order to avoid potential bugs between profiles not being detected until runtime. So profiles have been changed to be limited to deciding underlying database and switching functionality like swagger on. This commit fixes this and makes the H2 database the persistence layer in the default profile. Some structural changes happen as well to make things more sensible