Add override of equals method.
Coverity is reporting that we should override the equals method in all
our domain subclasses. This hasn't been an issue for us as the parent
class has en equals method and systemId should be unique and enough to
aid equals. This could perhaps be an issue if we clone an entity and
change some of its contents. But then the systemId should also be
changed. We are adopting equals to all domain model subclasses to be
more in compliance with best practice and aid perhaps situations where
part of the code-base is being used in other scenarios.