Remove the import API.
This code has some bugs in it reported by Coverity. Nothing serious. The
code isn't in use nor are we maintaining it. So there really is no point
in it being here anymore. It's in the git history so if we decide to
work with it again, we can pull it out from here.