Add missing methods.
Coverity is reporting that these methods are missing equals and will
then ask for hashCode. However I'm not convinced that this is the proper
solution. Perhaps we should be pulling out the instances of classes and
doing hascode and equals on these.

An own commit to show that I am aware of the issue.