Add missing methods.
Coverity is reporting that this class is missing equals and will
then ask for hashCode. However I'm not convinced that this is the proper
solution. Perhaps we should be pulling out the instances of classes and
doing hashcode and equals on these.

An own commit to show that I am aware of the issue.