Refactor Hateoas link creation.
This [issue](https://github.com/HiOA-ABI/nikita-noark5-core/issues/105)
has been a problem as we have been doing a lot of strange casting
between Set and List in the codebase. At some point we didn't have time
to fix it so we ended up creating an ArrayList from a Set, something
that really was ridiculous. We had time to look at the issue now and
are using AbstractCollection and that seems to be OK.

However runtest is reporting 7 new unexpected errors that are related to
correposndancepart. Commiting this anyway as it's such a big change.