Another attempt to refactor Hateoas link creation.
This [issue](https://github.com/HiOA-ABI/nikita-noark5-core/issues/105)
has been a problem as we have been doing a lot of strange casting
between Set and List in the codebase.

There are some prior commits that attempted to fix this but failed due
to mismatches between Hibernate types and more generic casts. The
approach now is to just make everything use a List and that seems to be
working. As far as I remember, we do not do any look ups in any lists.
All results are processed serially so this change should to Lists should
not be a problem.

Note, we should do a manual inspection of the code to make sure we do
no list based lookups (performance issue) and we should check that all
casts are correct. This can be done when we move the hateoas stuff to
the service layer.