Move Hateaos links generation from Controller layer to Service layer.
We need to get the generation of Hateaos links to occur within a
transaction boundary and are slowly going to move the functionality to
the service layer. This is the first attempt at doing this for the
Fonds 'stack' within nikita. There are various things happening during
this commit. It was difficult to find any point to push earlier as the
codebase was in a non-compilable state until a lot of changes were in
place.