Temporarily add OData parsing and ability to run simple OData parsing from OData $filter= .... to SQL.
Adding this to get all further OData parsing under revision control.
When we finish OData parsing it will probably be part of the util
package of core-common.

It should be possible to run TestODataApp as a java application and you
will see a filter command converted to SQL. TestODataApp will try to
open the odata_samples.txt and parse the URL and convert the filter part
to a where clause. Going forward this will convert to HQL rather than
SQL.