Update OData processing.
Finally got a decent structure on how to handle OData parsing, allowing
for parsing to SQL, HQL and possibly elasticsearch syntax. This
introduces OData contains and startsWith, both being translated to a
type of LIKE expression.

This also introduces mapping from Norwegian entity / attribute names
to English database tables/attributes fro SQL and class
objects/variables for HQL. ES is currently just using the HQL approach.

We are not really testing the ES approach. We're just adding code that
seems reasonable to identify if we have to handle things differently
with ES or not.