Make it possible to switch swagger on as a profile.
Swagger has been disabled for a while now. This is because it hangs
after printing the following message:

ApiListingReferenceScanner : Scanning for api listing references

I'm not sure why this is happening, but I think it's going into some
kind of loop and I'm not sure why. Adding @ApiIgnore to everything
except FondsController helped, but it hung at some point after that.

We really would like to see swagger support enabled, either statically
or dynamically, but we just don't have the time to figure it out. This
commit just ensure it's possible to start nikita with the swagger
profile enabled.

     -Dspring.profiles.active=swagger