Major structural change from multi-module to single module maven project
We were never able to get spring working "just right" as a multi-module
project. We weren't for example able to get spring testing working
properly. We also missed out on a lot of springs autoconfiguration and
had a lot of annotations on things that spring takes care of.

spring now automagically configures most things.

We used this opportunity to change the package naming aswell.  The old
naming was no.arkivlab.hioa.nikita. This naming came from an initiative
to try to create a community for public archives software called
arkivlab. Nothing ever came of this initiative and the project is firmly
anchored at OsloMet (earlier HiOA) so the need for package naming to
include organisation is redundant. HiOA is now called OsloMet so
removing hioa from the codebase makes sense.

This commit leaves the codebase broken. This is because the metrics
annotations do not work. Something in the manual configuration of
entities was preventing this issue from appearing, but it has
manifested itself now.