spring-metrics problems on controllers
@Counted and @Timed can't be run at the same time. If we wish to do this
then we have add a unique name in the constructor of the annotations.

We knew of this problem, but for some reason it wasn't showing itself
with the oldproject structure. My guess is that the spring-metrics was
in fact not being started for some reason.

So we remove all @Timed and will revisit this as the project progresses.
spring-metrics has not been updated since 2015 so there might be
something else to use instead. We will check this out.

Also reenable import to database. Not sure why it was commented out
and set metrics time to every 5 minutes to avoid noise in console.

This commit should fix the build.