Introduce Integration testing to the codebase
This is the first step of getting a testing framework up and running.
Two tests are introduced here, login test and create a fonds. There
are some casting issues so we need to look at little more at this and
we need to be able to deal with the hateaos links. Currently they are
ignored during deserialisation.