Update project to spring-boot version 2.
To get this to work caused a lot of issues. This was a pain to get
working and lot's of different things were tried before finding out the
reason we couldn't run was that <start-class> in pom.xml was pointing to
the wrong class.

This commit is probably a very good example of a bad commit. There is so
much happening in this commit. Project structure changed, a lot of the
repository classes changed to include <Optional> etc. Bar the project
structure change most changes in code here are related to an upgrade to
spring-boot 2.

The structural change, moving core-webapp to parent was done as I
thought there might be som pom issues related to why we couldn't start
the embedded webserver. It had to be done, but I would have preferred it
to be it's own commit.

This commit marks the fact that we upgrade to spring-boot 2 and a change
in project structure from a multi-module maven project to a single
module maven project. The latter is done as we were having a lot of
problems with configuration and we simply were unable to get it right!

This commit builds but does not run. There is a problem with JWT. But we
will move to springs built in JWT support now.