Update pom to reflect latest structure and spring2
A lot of project dependencies overlapped between core-common and
core-webapp. When both modules were merged the overlap became redundant
and we were able to rely more on springs own starter for dependencies.

Also there was a lot of legacy and bad practices in the previous pom
that has been removed.